ECL Computation Correction and Code Optimization Togo 2 half to 20 min

+++ IFRS9_ECL.Models/Framework/LifetimeCollateral.cs
@@ -9,9 +9,9 @@ namespace IFRS9_ECL.Models.Framework
     public class LifetimeCollateral
     {
         public string ContractId { get; set; }
-        public int EirIndex { get; set; }
-        public int TtrMonths { get; set; }
-        public int ProjectionMonth { get; set; }
+        public long EirIndex { get; set; }
+        public long TtrMonths { get; set; }
+        public long ProjectionMonth { get; set; }
         public double ProjectionValue { get; set; }
     }
 }
